% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.dptee.R
\name{functions.dptee}
\alias{functions.dptee}
\alias{functions}
\alias{lagnmatrix}
\alias{trimrmatrix}
\alias{dimen}
\alias{gmm}
\alias{estimate}
\alias{makeIV}
\alias{makeMF}
\alias{makeW}
\title{dptee internal functions}
\usage{

lagnmatrix(n, t, x, p)

trimrmatrix(n, t, x, p)

dimen(x)

gmm(n, y, x, iv, w)

estimate(coef, q, eq, z2, z1, n, t, ml, dy, iv, con, xk, kxk, effects, tt, ktt, ki, display.dummies)

makeMF(eq, data, effects, normal.inst)

makeW(n, iv, u = NULL)
}
\description{
{Functions are: (1) lagnmatrix, (2) trimrmatrix, (3) dimen, (4) gmm, (5) estimate, (6) makeIV, (7) makeMF, and (8) makeW.}

Function (1): lagnmatrix(n, t, x, p)

Function (2): trimrmatrix(n, t, x, p)

Function (3): dimen(x)

Function (4): gmm(n, y, x, iv, w)

Function (5): estimate(coef, q, eq, z2, z1, n, t, ml, dy, iv, con, xk, kxk, effects, tt, ktt, ki, display.dummies)

Function (6): makeMF(eq, data, effects, normal.inst)

Function (7): makeW(n, iv, u = NULL)
}
\keyword{internal}
